Refactored significant poritions of the PDF generation stack in order to better support a generalized concept of grid measuring and rendering. This was inspired by a lack of support for multi-page pdfs in the update, so this is handled now. In order to acheive this in a general manner, I also upgraded the min SDK version to Lollipop

+++ smartReceiptsLibrary/src/main/java/co/smartreceipts/android/workers/reports/pdf/pdfbox/PdfBoxContext.java
@@ -45,9 +45,9 @@ public interface PdfBoxContext {
 
     int getLineSpacing();
 
-    int getPageMarginHorizontal();
+    float getPageMarginHorizontal();
 
-    int getPageMarginVertical();
+    float getPageMarginVertical();
 
     @NonNull
     String getString(@StringRes int resId, Object... args);
